backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-cors-headers==3.13.0
drf_yasg==1.21.7
psycopg2-binary==2.9.3
python-dotenv==1.0.0